requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anaconda-anon-usage @ file:///home/task_176160626771849/conda-bld/anaconda-anon-usage_1761606317374/work
anaconda-auth @ file:///opt/miniconda3/conda-bld/anaconda-cloud-auth-split_1759943045113/work
anaconda-cli-base @ file:///opt/miniconda3/conda-bld/anaconda-cli-base_1762294806728/work
annotated-types @ file:///opt/miniconda3/conda-bld/annotated-types_1761744949290/work
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
async-timeout==4.0.3
attrs==23.2.0
audioop-lts==0.2.2
beautifulsoup4==4.12.3
boltons @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_97h34kncb9/croot/boltons_1751383736162/work
brotlicffi @ file:///opt/miniconda3/conda-bld/brotlicffi_1761745164052/work
bs4==0.0.2
certifi==2024.6.2
cffi @ file:///opt/miniconda3/conda-bld/cffi_1761832773185/work
charset-normalizer==3.3.2
click @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d2rtg2ejqc/croot/click_1744271604580/work
conda @ file:///opt/miniconda3/conda-bld/conda_1760457876773/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ebkcs5yxh6/croot/conda-anaconda-tos_1755123330773/work
conda-content-trust @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-content-trust_1728603284288/work
conda-libmamba-solver @ file:///opt/miniconda3/conda-bld/conda-libmamba-solver_1761744809966/work/src
conda-package-handling @ file:///opt/miniconda3/conda-bld/conda-package-handling_1762366402479/work
conda_package_streaming @ file:///opt/miniconda3/conda-bld/conda-package-streaming_1762361629849/work
cryptography @ file:///opt/miniconda3/conda-bld/cryptography-split_1761931964095/work
discord==2.3.2
discord.py==2.6.4
distro @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/distro_1728591855492/work
exceptiongroup==1.2.1
frozendict @ file:///opt/miniconda3/conda-bld/frozendict_1761750665916/work
frozenlist==1.4.1
h11==0.14.0
idna==3.7
jaraco.classes @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_97dr4u6nc1/croot/jaraco.classes_1755516335437/work
jaraco.context @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jaraco.context_1731716684676/work
jaraco.functools @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_91zaq5_008/croot/jaraco.functools_1740408121405/work
jsonpatch @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jsonpatch_1728594556110/work
jsonpointer @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_bdid71leqg/croot/jsonpointer_1753788445166/work
keyring @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_be9at15ui5/croot/keyring_1745484704517/work
libmambapy @ file:///opt/miniconda3/conda-bld/mamba-split_1757622986409/work/libmambapy
lxml==5.2.2
markdown-it-py @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6av49xs_3l/croot/markdown-it-py_1756299201304/work
mdurl @ file:///opt/miniconda3/conda-bld/mdurl_1758552175942/work
menuinst @ file:///opt/miniconda3/conda-bld/menuinst_1762860695894/work
more-itertools @ file:///opt/miniconda3/conda-bld/more-itertools_1761121541654/work
multidict==6.0.5
numpy==1.26.4
outcome==1.3.0.post0
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079023/work
pandas==2.2.2
pkce @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pkce_1728592680499/work
platformdirs @ file:///opt/miniconda3/conda-bld/platformdirs_1762356666077/work
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_bei6q1cvvt/croot/pluggy_1733169621717/work
prompt_toolkit==3.0.52
propcache==0.4.1
pyaes==1.6.1
pyaml==24.4.0
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_61k1o_oje7/croot/pycosat_1736868429741/work
pycparser @ file:///opt/miniconda3/conda-bld/pycparser_1757496048647/work
pydantic @ file:///opt/miniconda3/conda-bld/pydantic_1762505197323/work
pydantic-settings @ file:///opt/miniconda3/conda-bld/pydantic-settings_1758784213628/work
pydantic_core @ file:///opt/miniconda3/conda-bld/pydantic-core_1760597286482/work
Pygments @ file:///opt/miniconda3/conda-bld/pygments_1762431584667/work
PyJWT @ file:///Users/builder/cbouss/buildout/croot/pyjwt_1739484683708/work
Pyrogram==2.0.106
PySocks @ file:///opt/miniconda3/conda-bld/pysocks_1761753010927/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0bqseadoho/croot/python-dotenv_1745610246418/work
pytz==2024.1
PyYAML==6.0.1
questionary==2.1.1
readchar @ file:///opt/miniconda3/conda-bld/readchar_1760613300945/work
requests==2.32.3
rich @ file:///opt/miniconda3/conda-bld/rich_1760375480293/work
ruamel.yaml @ file:///opt/miniconda3/conda-bld/ruamel.yaml_1762535951282/work
ruamel.yaml.clib @ file:///opt/miniconda3/conda-bld/ruamel.yaml.clib_1762530070208/work
selenium==4.21.0
selenium-stealth==1.0.6
semver @ file:///opt/miniconda3/conda-bld/semver_1761903026306/work
setuptools==70.0.0
shellingham @ file:///opt/miniconda3/conda-bld/shellingham_1761912204150/work
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
TgCrypto==1.2.5
tomli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03b1up8uh3/croot/tomli_1753774598375/work
tqdm @ file:///opt/miniconda3/conda-bld/tqdm_1762863348900/work
trio==0.25.1
trio-websocket==0.11.1
truststore @ file:///opt/miniconda3/conda-bld/truststore_1762520923820/work
typer @ file:///opt/miniconda3/conda-bld/typer_1757580868385/work
typing-inspection @ file:///opt/miniconda3/conda-bld/typing-inspection_1760614169744/work
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.14
wheel==0.45.1
wsproto==1.2.0
yarl==1.22.0
zstandard @ file:///opt/miniconda3/conda-bld/zstandard_1758189078532/work
